fix unit tests, seems to work now

+++ src/test/kotlin/renegade/aggregators/DoubleAggregatorSpec.kt
@@ -1,7 +1,7 @@
 package renegade.aggregators
 
-import io.kotlintest.matchers.shouldBe
-import io.kotlintest.specs.FreeSpec
+import io.kotest.core.spec.style.FreeSpec
+import io.kotest.matchers.shouldBe
 import renegade.approx
 import renegade.util.math.stats.count
 
@@ -13,7 +13,7 @@ class DoubleAggregatorSpec : FreeSpec() {
             "with values 2.0 and 2.1 with distances 0.1 and 0.2 respectively" - {
                 listOf(ItemWithDistance(2.0, 0.1), ItemWithDistance(2.1, 0.2)).forEach { agg.addValue(it) }
                 "verify they are aggregated correctly" {
-                    agg.summary.count shouldBe approx(2.0)
+                    agg.summary.count.toDouble() shouldBe approx(2.0)
                     agg.summary.sum shouldBe approx(4.1)
                     agg.summary.mean shouldBe approx(2.05)
                 }
